2026-14-02: Push general logic out of service, Prepared an interface
