+++ backend/src/main/java/com/localcode/services/CodeHarness.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.localcode.dto.ExecutionRequest;
-import com.localcode.dto.ProblemDetailDTO;
+
 import com.localcode.services.Emitters.CodeEmitter;
 import com.localcode.services.Emitters.EmitterFactory;
 
@@ -26,7 +26,7 @@ public class CodeHarness {
         CodeEmitter emitter = emitterFactory.getEmitter(request.getLanguage());
         StringBuilder harness = new StringBuilder();
 
-        harness.append(emitter.generateImports());
+      
         harness.append(emitter.generateTailCode(request.getMethodToCall()));
 
         return harness.toString();
