+++ backend/src/main/java/com/localcode/services/Emitters/CodeEmitter.java
@@ -2,9 +2,7 @@ package com.localcode.services.Emitters;
 import com.localcode.services.DataType;
 
 public interface CodeEmitter {
-    String generateImports();
+  
     String generateTailCode(String methodToCall);
-    String generateInputParsing(DataType dataType);
-    DataType dataTypeMap(String paramType);
 
 }
