+++ backend/src/main/java/com/localcode/services/TailCodeGenerationStrategy.java
@@ -1,0 +1,10 @@
+package com.localcode.services;
+import java.util.List;
+
+public interface TailCodeGenerationStrategy {
+    
+    String addDeclarations(List<Param> params);
+    String addCustomDataTypeClasses(List<Param> params);
+    String generateMethodCall(String methodName, List<Param> params);
+
+}
